+++ tests/test_getexampledata.py
@@ -1,0 +1,3 @@
+from pyraingen.get_example_data import get_example_data
+
+get_example_data(daily=True, subdaily=True, ifd=True)