+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Changelog
 
-<!--next-version-placeholder-->
+## v1.0.2 (18/12/2022)
+
+- Fix bug in get get_example_data.py
+
+## v1.0.1 (17/12/2022)
+
+- set dependencies in pyproject.toml as static 
+
+## v1.0.0 (16/12/2022)
+
+- Major update to regionalised daily sim nearby site selection.
+- Minor bug fixes
+- Updated example
+- Updated example data
+
+## v0.2.3 (01/12/2022)
+
+- ????
 
 ## v0.2.2 (30/11/2022)